`Debug` for `config::Snapshot`. (#331)
It's a bit special as it works around the release-versions inability
to reproduce itself losslessly. In debug mode, we do load files
withtout loss, making nice debug printing possible.
